// frontend/src/api/model/githubCom4H1RZooraInternalDomainUserConnector.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainConnectorType } from './githubCom4H1RZooraInternalDomainConnectorType';

export interface GithubCom4H1RZooraInternalDomainUserConnector {
  created_at?: string;
  enabled?: boolean;
  id?: string;
  target?: string;
  type?: GithubCom4H1RZooraInternalDomainConnectorType;
  updated_at?: string;
  user_id?: string;
  verified_at?: string;
}