// frontend/src/api/model/githubCom4H1RZooraInternalDomainConnectorType.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GithubCom4H1RZooraInternalDomainConnectorType = typeof GithubCom4H1RZooraInternalDomainConnectorType[keyof typeof GithubCom4H1RZooraInternalDomainConnectorType];


export const GithubCom4H1RZooraInternalDomainConnectorType = {
  ConnectorTelegram: 'telegram',
  ConnectorBale: 'bale',
  ConnectorSMS: 'sms',
  ConnectorPush: 'push',
} as const;